Reduce the minimum CPU interval
On some newer HW the current value is too high. This ends up causing
test failures. In the future the code should be investigated to see if
we could remove this value altogether but for now I am just trying to
stabilize testing.

In the future this code should also be investigated to remove the
NO_HISTORY option. This could likely be removed by caching the CPU times
on startup of the library.

Fixes #852

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>