Restore TR::CFGEdgeList performance
In 13e5f333, the container used by CFG nodes to store pointers to their
incoming / outgoing CFG edges was changed from ::List to, essentially,
an std::list. From a software-engineering perspective, this was quite an
accomplishment, changing around three thousand lines that interacted
with the container type while providing an interface idiomatic with
respect to the C++ standard library. On a performance front, however,
the change caused a regression, increasing the amount of CPU consumed by
the JIT compiler to perform equivalent tasks.

In order to restore the performance of the JIT compiler, this change
does a number of different things:

1) This change modifies the container type each CFG node's lists of
edges to use a forward_list.

2) Because we need to build the JIT compiler on platforms with no C++11
support beyond decltype and auto, this change provides a partial imple-
mentation of forward_list sufficient for the purposes of CFG nodes. We
also include a size() function that runs in time linear to the elements
in the forward_list that is supplemental to the standard's definition.

3) As a result of 1, the action of conditionally finding a value and re-
moving it from the container now requires 2 iterators rather than 1, as
forward_list implements erase_after rather than erase.  This change
updates the locations affected by this.

4) This change modifies the behaviour of TR_CFGIterator objects such
that they use references to the lists of edges they are iterating
through, rather than localized copies.

5) We redefine the type used by TR::ThreadLocalAllocator, eliminating
the use of CS2::stat_allocator. In the compiler's current form, the
stat_allocator functionality was impossible to enable without a rebuild.
However, it contributed a run-time check [always false] and an increase
in instruction length to allocation operations that went through our
TR::Allocator.

With these changes, on the original system that observed the regression,
the regression is reversed.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>