Use `extern` to defer function resolving to the linker
The function `void ppcCodeSync(uint8_t *, uint32_t)` was being declared
as `extern` in compiler/runtime/Trampoline.cpp only if the host
architecture was power. This resulted in errors during cross
compilation. This fix ensures that the above function is declared as
`extern` on either case: the host being power, the target being power,
or both being power.

Signed-off-by: Sajid Ahmed <Syed.Sajid.Ahmed23@ibm.com>