Check the number of pending pushes on a method
Under OSR, it will not be possible to transition correctly
if we allocate more pending push slots than the VM expects. To ensure
this never occurs, the JIT will perform a fatal assertion if it
detects a requested slot beyond the limit.

To assist in preventing this from occuring, ByteCodeIteratorWithState
will determine if OSR is possible within a method. Pending push
stores can then be avoided if the stack were to not match the VM's
expectation.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>