Fix more warnings in compiler component
Eliminates many common compiler warnings during build on OS X, at least.

Boiled down to about 7 issues. Only two of which had functional aspects,
I believe: 1) there was a recognized method test against -1, but the
"unknownMethod" value is 0, and 2) callerIndex was being asserted to be
of the right size so that upper bits could be properly stolen, but it
was using INT_MAX in the calculation whereas the caller index is a short.

Otherwise, this commit is just moving text around or adjusting printf
formats to eliminate warnings.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>