Do not decrease ref count of compress sequence in l2aEvaluator
Previously we used to evaluate the shift left child when
l2a node contains compression sequence under Concurrent
Scavenge mode. As this part of code was refactored, we
do not need to decrement reference count of the shift left
child of the tree as we don't evaluate it in this function.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>