Add OSR point for NextGenHCR in outermost method
Under NextGenHCR, it may be necessary to transition from
the start of the method, in the event that an assumption
was invalidated during the method prologue. A prior change
added exceptions edges to handle this, however, parts
of the OSR API require an OSR point to also be present. This
change adds the OSR point, with the correct bytecode index.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>