Fix double 0x in log file pointers on OSX
The nest of conditionals in env/OMRIO.hpp that determines what
printf format incantation to use for pointers (i.e. that defines
POINTER_PRINTF_FORMAT) ends up in a strange branch for OS X
and that means pointers get an annoying 0x0x prefix in the
trees that makes it harder than required to get around.

Treating OSX like Linux does the trick.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>