Cheaper bump instructions for small frames on x86
We typically generate a SUB/ADD instruction to buy the stack frame
at the beginning/end of methods. This particular flavour of SUB/ADD
consumes 4 bytes. If the stack frame is very small (say 8 bytes
on 64-bit or 4 bytes on 32-bit) then the frame can be bought
with a simple PUSH/POP operation.

This commit just special cases this small frame size and uses
PUSH rbx / POP rbx (on 64-bit, ebx on 32-bit).  Note that *bx
is callee preserved so it should be a reasonable choice.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>