Implement isLeafMethod for x86 code generator
To simplify frame prolog/epilog in small methods, it is handy to
know if anything will be called. This commit introduces a flag
on OMRCodeGenerator that captures whether the method being
compiled is a leaf method or not. The flag is initialized to
true in the constructor. Every spot in the x86 code generator
where a CALL instruction is generated, this flag is reset.

Nothing uses this flag yet; that will come in a separate commit.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>