Do not allocate args space for leaf methods on X86
For AMD64, we try to 16-byte align the entry stack pointer for
called methods by allocating space in the caller's frame to
compensate for the return address pushed on the stack by a call.

For leaf methods, there will be no calls, and so there's no
reason to consider 8 bytes being added for a return address
and then to align that 8 bytes to 16 bytes. This commit
uses the fact that the current method is a leaf method to
avoid adding 16 bytes to the stack pointer, which in some
small methods means no stack frame is actually needed.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>