Implement omitFramePointer for X86 system linkage
That pesky frame pointer isn't always useful and, in fact,
can hurt performance if prologs and epilogs are common
(i.e. you're doing tons of calls). A recent peformance
analysis identified the presence of the frame pointer as
a performance gap, so this commit contains the code to
make the use of a frame pointer selectable via an option
-Xjit:omitFramePointer .

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>