Update ftops.pl and associated files
The 'ftops.pl' tool is used to generate
parts of the Power code generator, however
it must (currently) be manually executed
and the resulting artifacts must be checked in.

This patch cleans up the code to make it
easier for developers to run and maintain,
re-generates the artifacts, and removes
dead artifacts and code.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>