Fix for unsafe generation of VFMA and VFMS on z
generateFusedMultiplyAddIfPossible tries to check if it is working on
floating point data before generating fused instructions. The problem is
caused by the fact that a call to isFloatingPoint on a type returns false
for VectorDouble and VectorFloat types. To address this issue, a specific
check for VectorDouble and VectorFloat was added to the function
generateFusedMultiplyAddIfPossible.

Issue: #1160
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>