Fix an "undefined symbol" runtime error on zLinux
The template method `loadAndStoreMayOverlap` in `OMR::Z::CodeGenerator`
was implemented in the implementation file rather than the header file,
and, hence, was violating the C++ standard. GCC 4.8.5 is inlining the
template straight into the overload candidate and not emitting the
template instantiation. The result was an "undefined symbol" error
during runtime on zLinux.
Issue: #1183

Signed-off-by: Alexey Zhikhartsev <alexey.zhikhartsev@ibm.com>