Try removing if node in all compare simplifiers
Having an if node with its branch target the same as its immediate fall
through block results in an illy formed cfg that will confuse other
optimizations. Removing such node is supported in most comparison
simplifiers but is missing in ifacmpne and ifacmpeq, and in generic
simplifiers that handles if nodes such as dftSimplifier. This commit
adds code to try to remove such if node in the simplifier.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>