Fix GCExtensionsBase initialization
To construct GCExtensionsBase, we use memset, then placement new.
Placement new is considered to fully initialize all memory of the
object. This causes the compiler to treat the memset as a dead store.
Some compilers will completely omit the memset, and may cause some
fields of GCExtensionsBase to be indeterminately initialized. The
solution is to fully initialize GCExtensionsBase in it’s constructor,
and stop calling memset.

Closes #1220
signed-off-by: andrew young <youngar17@gmail.com>