z/TPF OMR Port Pull Request 2
This pull request is the second in a series of committing omr changes
for the z/TPF platform. This pull request is a follow-on to PR #1147.
A successful personal build was done in the IBM environment
(build_id=351866), there were two or three compilation failures that
were fixed. Unlike PR #1147 there are a small amount of platform
common files which were updated.

It's expected there will be two more pull requests after this one of
smaller sizes then followed by pull requests for specific fixes as part
of z/TPF builds and runtime fixes.

The current z/TPF build procedures are still a work-in-progress.

Most of the formatting was pre-existing. Some of the files maybe
removed once our build is finalized. So no need to overly analyze this
pull requests as we still have passes coming as we work out our build
and runtime bugs.

Signed-off-by: jjohnst-us <jjohnst@us.ibm.com>