Create VMMethod extensible class
Create an interface to a VM to ask questions about methods.  This class
will become the home for current FrontEnd functions that accept
`TR_OpaqueMethodBlock`s as their primary parameter, and in most cases
simply delegate to a VM.  However, the presence of such a parameter
is not necessarily a requirement for a function to be in this class.

The starter set of functions are lifted (but not yet removed) from
FrontEnd.hpp:

* hasBackwardBranches
* isCompiledMethod
* getMethodStartPC

An instance of this class is added to the CompilerEnv structure.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>