Flush verbose GC buffer every heap-resize
Previously, heap-resize operation is subtask of gc-cycle. So, it was
not required to flush verbose GC buffer immediately. But, as part of
the EXPLICIT_GC_IDLE_GC, heap-resize type="release free pages" may be
executed outside the gc-cycle. For those cases, the immediate buffer
flush required for the tag appear in the verbose output.

Signed-off-by: Parameswaran Selvam <parselva@in.ibm.com>