Replace redundant lwz with rlwinm
Power processors aren't optimized for storing values and then fetching
them back, so the compiler optimizes away redundant loads after
stores. In some cases, lwz was used to clear the high 32 bits of a
variable, but the compiler optimized it away regardless. Replacing it
with rlwinm, which will clear the high 32 bits solves this.

Signed-off-by: Agoston Szepessy <agoston.szepessy@ibm.com>