Avoid insert asyncchk in arraytranslate opt
The OSR transition failed because of an asyncchk inserted
by arraytranslate. The asyncchk is only needed if asyncchk
removal might have removed an original asynchk in the containing
loop. Asyncchk removal is disabled if we are in either one of
involuntaryOSR or nextGenHCR modes, and therefore the inserted
asyncchk should be avoided under these 2 modes.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>