Fix redundantGotoElim asynccheck insert for OSR
Changes were made to redundantGotoElim to allow it
to perform under NextGenHCR without placing an
asynccheck where it may invalidate an assumption
without being protected by an OSR guard.

Incorrect changes were made when adding the possibility
of both pre and post execution in one compile.
Moreover, the modification to the optimization is
specific to how assumptions are made, not the OSR
mode.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>