Add tests for the Tree Verifier
The Tree Verifier is supposed to check that IL is well-formed and follows
the rules of Testarossa's intermediate language. The tests added by this
commit check that the verifier rejects ill-formed IL and accepts
well-formed IL.

Some of these tests currently fail as the verifier has not been
historically well maintained.

[ci skip] because Tril is not part of continuous integration

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>