Run the unimportantLoopCount at all optlevels
In TR_LoopVersioner::loopIsWorthVersioning we currently only
run the test for unimportantLoopCount for optlevels <= warm. It is
likely to be beneficial in many circumstances to run this test at
all optlevels.

This change is currently guarded by the Xjit option added in
PR #1369 until such time as the behaviour is tested accross a broad
set of applications.

Signed-off-by: JamesKingdon <jkingdon@ca.ibm.com>