RAS capabilities for compiler scratch memory
This commit provides two new capabilities for interacting with scratch
memory:

1) The RegionProfiler class is intended for tracking changes in Region
memory consumption between the construction and destruction of the
RegionProfiler.

2) The DebugSegmentProvider class provides an aggressive failure mode
used to detect invalid reads and writes to expired memory regions. This
is achieved by changing the memory mapping of a memory segment when it
is relinquished by a Region back to the SegmentProvider, freeing the
physical pages and marking the mapping as having an empty permission
set.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>