Check return temp exists splitting in inlining
Under OSR, its necessary to ensure the last block
of the callee and the caller's block immediately
after must be separated for distinct OSR exception
edges. However, this splitting was not taking into
account whether the return value was commoned across.
This was working in the past because the callee
always had an OSR catch block, resulting in
the inliner generating a temporary for the result.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>