Compilation object cleanup
Change the Compilation object's region member type to a reference,
allowing comp->region() to return the same region used by heap memory.

Clean up the initializer list of OMR::Compilation so that members
initialization are specified in the same order as their declaration
inside the class definition. (Which is the order they get initialized
anyway, regardless of their order in the initializer list.)

Change permissions of _signature member and replace its external uses
with a call to the signature() method. _signature should still be the
first member, as this can significantly ease defect investigations.

Remove _aotClassInfo from OMR::Compilation as it never gets used.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>