Store value to auto in IlValue::storeOver
IlValue::storeOver does not handle the incoming parameter "value"
properly, in that it does not ensure that it has been stored to
an auto symbol if it's computed in a different block. This mistake
can result in illegal IL in the form of node references that cross
basic block boundaries.

The fix ensures that value has been written to an auto if necessary,
and also leaves the IlValue in a more consistent state (earlier
version would update _nodeThatComputesValue without also updating
_blockThatComputesValue; really neither of these should be updated).

Fixes: #1305

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>