Fix live monitor meta data to detect invalid cases
In original lmmd code, when propagating monitor state from one block
to its sucessors the blocks that are already visited are skipped
and it's assumed that the monitor state for one block should be the
same no matter which predecessor it's propagated from.
This code detect the cases where the monitor state for one block can vary
when propagating from different predecessor blocks. In those cases,
lmmd should resort to the most conservative solution.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>