Compress OSR infra after all OSR points are added
OSR slot sharing information is intended to be
compressed after the map from instruction to slot
information has been generated. However, it is
being compressed before the slot sharing information
for snippets has been added. This can result in
a transition failing if a snippet's instruction PC
is within a range of compressed PCs and has different
slot sharing information.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>