Remove the _opt field in ILWalk and ILValicator
The IL iterators previously had an `_opt` field of type
`TR::Optimization`. This dependency is undesired as IL iterators should
be usable outside of the optimizer. Since the object is only used to
gain access to the compilation object, this change replaces the `_opt`
field with the `_comp` field and makes the dependence on the compilation
object direct. In addition, the ILValidator, which uses an IL iterator,
is also updated to depend directly on the compilation object instead of
an optimization.

Also, the `_comp` parameter is no longer optional and does not default
to NULL. This is because it does not currently make sense to use any of
these APIs outside the scope of a compilation, since the IL is only
available during compilation.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>