Minor ILValidator fixes
This commit does two things: it changes a `NULL` to a `0` and inverts a
validation condition.

Changing the NULL to a `0` in one of the ternary expressions ensures
that the types are consistent. This prevents the flowing gcc warning:
"converting to non-pointer type ‘long int’ from NULL".

The condition for checking if the end of an extended block has been
reached by the validator was corrected by inverting it. The original
condition stated the current node is the end of an extended block if the
block that follows *is* an extension. This is clearly incorrect because
if the following block is an extension, then we haven't reached the end
yet. The fix is to invert the condition.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>