Tracegen tool fixes
Tracepoints/TraceAssert normally have a default parameter which is
the current thread.  There is an option NoEnv to remove the default
current thread parameter.  The template for generating the TraceAssert
was not taking the envParam into consideration.  This means that
TraceAsserts which did not use NoEnv were generating incorrect data.
The fix is to take it into consideration and generate the correct code
for these asserts.

Header file creation now throws an error if the Trace point is not
properly formatted rather than being generated and then later causing
a compilation error.

Signed-off-by: Lina Serry <linaserry@cmail.carleton.ca>