Deduplicate compiler build support scripts
When the compiler makefiles were redone (prior to open-sourcing OMR)
there was duplication injected into the Makefiles to provide maximum
flexibility across projects with different build needs.

One place where we can now start to reduce that build flexibility is the
build support scripts. This commit moves the support scripts to a common
location and starts using them for the OMR contained compiler projects
(jitbuilder and the test compiler).

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>