Fix errors found by validator in JitBuilder
The only problem reported by the ILValidator in JitBuilder has to do
with the implementation of VirtualMachineRegister and OperandStack.
Essentially, the pointer to the base of the stack was being stored in an
IlValue with the type of the elements of the stack (instead of an
address). This resulted in ill-formed tree being generated and reported
by the ILValidator.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>