Allocate using regions in TR_List and TR_BitVector
The current implementations of TR_BitVector and TR_List allocate memory
using trMemory(). This is problematic for growable data structures using
stack memory since new stack marks may cause the lifetime of internal
data structures to be inconsistent. In addition, the lack of a
TR::Region based API prohibits modelling memory lifetimes which do not
follow the existing stack mark or heap.

This change modifies the data structures to store a region to do their
memory management and to always use the same region for their
allocation. The trMemory based APIs derrive the appropriate default
region to minimize impact on the code.

To facilitate allocation from region overrides of the pacement new and
delete operators taking TR:Region are supplied to mirror those for
trMemory and friends.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>