Allocate region for AuxiliaryData in UseDef
Currently the UseDef analysis has an auxiliary structure used to hold
data only live during usedef analysis. This structure, while allocated o
the stack, is making heavy use of the CS2 allocator. Much of the heap
memory used is not reclaimed and is leading to large footprint
overheads.

This change modifies TR::deque to allow a non-CS2 allocator to be
supplied and uses this deque along with a region embeeded in the
auxiliary usedef structure to allow for efficient memory clean-up.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>