Remove CS2::ArrayOf from UseDef auxiliary struct
The UseDef auxiliary strucutre is currently making use of the
CS2::ArrayOf container which is not compatible with region allocation.
This change introduces a TR::vector similar to the TR::deque and uses
this data type to replace the ArrayOf in UseDef's auxiliary strucutre.
This allows the arrays to be reclaimed efficiently.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>