Region to allocate structures for UseDef proper
Currently the data structures populated by UseDef analysis and utilized
by the callers are allocated on the heap making heavy use of CS2 and its
allocator. There are significant memory leaks from this. This change
introduces a region for UseDefInfo and allocates the structures
populated from this region. As part of this TR::list is modified to
allow a non-CS2 allocator to be supplied so that it is compatibile with
region allocation.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>