Remove CS2 BitVectors from DataFlow
The DataFlow engine was designed to operate with stack memory and the
associated stack marks and releases. The BitVectors were previously
changed to use CS2 dense BitVectors allocated out of a CS2::TableOf
pool. The pool is leaking large amounts of memory and the CS2 dense
BitVectors are not compatible with region allocation. This change swaps
the BitVectors back to region allocated TR_BitVectors and removes the
CS2::TableOf pool for significant memory savings.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>