Clean RegisterCandidates TableOf refAutos
RegisterCandidates currently uses a CS2::TableOf to track referenced
autos as part of its analysis. This strucutre is leaking a lot of memory
so this change converts it to use an std::map that is region allocated.
As part of the region allocation this change converts some CS2
BitVectors and SparseBitVectors to TR_BitVectors. Some unused fields are
also cleaned up.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>