Cleanup RegisterCandidates CS2::TableOf
The RegisterCandidates structure contains a CS2::TableOf holding
allocated candidates that is leaking large amounts of memory. This
change replaces it with a region allocated std::map. To allow the data
in the map to be region allocated backing data structures are clean-up
in some of the utility data structures.

There is a second TableOf which holds Coordinate information for a
reference, this table is also cleaned up using std::map and region
allocation.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>