Clean Dominators memory management
Currently a Dominators analysis has a number of data structures scoped
to the lifetime of the analysis. These structures are currently CS2
containers allocated using the CS2 allocator. This change refactors
Dominators to have an embedded region and switches the data structures
to all allocate from that region. The TableOf in the current
implementation was especially bad at leaking memory.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>