Remove CS2::TableOf from CFG
The CFG current has a CS2::TableOf which is used to allocate a 'globally
unique ID'. This ID is only used as part of the RegionStructure to track
its child elements and search for cycles. These functions can be
completely implemented using the CFG Node IDs and so we can remove the
TableOf and change the RegionStructure to store its nodes as a
TR::vector simplifying many operations on the structure.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>