Add Tril-based tests for shift and rotate opcodes
These tests will replace the shift and rotate opcode tests that are
currently in the test compiler. The new tests correctly use an Int32
for the shift amount used to test all opcodes, unlike the old ones
that assumed the shift amount must be the same type as the value
being shifted.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>