Create a memory region for structure nodes in CFG
Currently structure is allocated on the heap and when structure is
invalidated the pointer to the root of the structure is simply NULL'd
and the structure objects leak. This change creates a region owned by
the CFG into which structure nodes are allocated. Once additional
clean-up is completed, the goal is to destroy and create the structure
region in place when structure is killed to release the memory rather
than leaking it.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>