Move BlockStructure flags to block
There are currently flags on BlockStructure which are copied forward
from one generation of structure to the next. This involves looking at
old structure even after it is invalidated. Since the properties copy
forward, we move these flags to the block and make the BlockStructure
API a passthrough so that we do not need to look at stale structure
information.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>