Stop converting references to integers
Converting object references to integers would ruin the collectable
property on a node when the address is a reference to a collectable
object, resulting in gc crashes. This commit removes conversions from
address to integer and add support to address comparisons. Changes
include

1) Remove a2l/a2i in LoopVersioner in LoopVersioner, LoopStrider,
GeneralLoopUnroller, InductionVariableAnalysis and ValuePropagation.
2) Create primary induction variable for some address types when they
are associated with an internal pointer.
3) Add UnsignedCompare and Unsigned data type for address compare
opcodes and update their VPHandlers.

Signed-off-by: liqunl <liqunl@ca.ibm.com>