Add two queries to answer node collectedness
The query isArrayRef by its name is supposed to return true for nodes
who are addresses pointing to the middle of objects. However, it is a
query on an opcode and returns true for aiadd and aladd, which is
insufficient to serve its original purpose. In some places in the code,
the isInternalPtr flag on a node will be incorrectly set if isArrayRef
returns true, resulting crashes in the gc when it tries to scan the
object the node points to. Unfortunately, isArrayRef is used everywhere
to check if the opcode is aiadd or aladd or to serve other purposes,
thus we can not simply change the implementation of this query or
replace it everywhere with a different query.

This commit includes the following changes:
1) Add query computeIsCollectedReference
2) Add query computeIsInternalPointer which makes use of
   computeIsCollectedReference
3) Modify the existing query isNotCollected to use the more accurate
   query computeIsCollectedReference
4) Replace the use of isArrayRef with computeIsInternalPointer where
   isInternalPtr flag might be set when isArrayRef returns true.

issue: #563
Signed-off-by: liqunl <liqunl@ca.ibm.com>