Support building compiler components with CMake
This adds support for building compiler components with CMake, and
demonstrates with the ability to build and run certain JitBuilder
examples.

The functionality to build the compiler is encapsulated in a CMake
function called create_omr_compiler_library. This function does the
heavy lifting to allow someone to describe a new compiler library with
relative ease.

This commit is part of the experimental CMake support, and so, the CMake
support isn't completed by this commit. Limitations include:

 - The support is all hardcoding x86-64 properties. Platform
   independence will come in subsequent commits.
 - The JitBuiler release directory is not populated, so the JitBuilder
   examples have access to more of the compiler interface than is
   intended.
 - The macro used to describe the JitBuilder examples doesn't have a
   sufficiently rich interface to describe how Mandelbrot is invoked.
 - Not all examples are built and run
 - The JitBuilder tests are not hooked up here.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>