CMake: use `-pthread` on linux and osx
Use `-pthread` as a compile flag for all objects. This causes gcc to
define the `_REENTRANT` flag, so do not define it ourselves.  We don't
directly define `_REENTRANT` because it is a compiler detail.

Use `-pthread` when linking against `j9thrstatic`. This was the
forgotten step in #1352 that caused missing pthread symobls.

Signed-off-by: Robert Young <rwy0717@gmail.com>