Create a region to hold alias TR_BitVectors
The current aliasing infrastructure allocates bit vectors to return to
callers of the aliasing interface. These bit vectors are heap allocated
and leak to the end of the compile. This change introduces a region to
hold these bit vectors and an API on the compilation to signal when it
is safe to invalidate the region and release the memory. The change adds
a call in LocalCSE to this API as well as calling it between
optimizations to reduce memory usage.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>