Remove optDetailString from OptimizationManager
Currently the optimization details string used during tracing is
initialized as part of the optimization manager construction. Because
OMR compiler consumers may initialize some optimizations themselves,
this opens us up to a scenario where these strings start to diverge
between the various consuming projects.

As developers rely on the consistency of those strings in order to aid
in cross-project debugging, this change modifies these optimization
string definitions such that they are an inherent part of the class
definition, and are accessed via a virtual function, for which each
implementation returns a statically defined string.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>