Modify NodePool to use a Region allocator
The current NodePool uses a CS2 TableOf as its allocator leading to some
strange code convetions when it comes to allocating nodes, recreating
nodes, and otherwise handling node lifetimes. This commit changes the
pool to use its own region and modifies the Node implementation to be
less dependent on the NodePool backing allocator.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>